clean up stepForward action sequence
