--- src/js/redux/actionSequences/tutorial/handleNextTutorialStep.js
@@ -1,7 +1,0 @@
-import handleSubStep from './handleSubStep.js'
-
-export default () => (dispatch, getState) => {
-  const { updating } = getState().tutorial
-
-  if (!updating) return dispatch(handleSubStep())
-}
